+++ src/main/kotlin/site/siredvin/peripheralium/api/blocks/IBlockObservingTileEntity.kt
@@ -1,0 +1,11 @@
+package site.siredvin.peripheralium.api.blocks
+
+import net.minecraft.core.BlockPos
+
+interface IBlockObservingTileEntity {
+    fun placed() {}
+    fun destroy() {}
+    fun onNeighbourChange(neighbour: BlockPos) {}
+    fun onNeighbourTileEntityChange(neighbour: BlockPos) {}
+    fun blockTick() {}
+}