+++ src/main/kotlin/site/siredvin/peripheralium/api/peripheral/IPeripheralTileEntity.kt
@@ -9,5 +9,5 @@ import net.minecraft.world.level.block.state.BlockState
 interface IPeripheralTileEntity {
     val peripheralSettings: CompoundTag
     fun markSettingsChanged()
-    fun <T : BlockEntity?> handleTick(level: Level?, state: BlockState?, type: BlockEntityType<T>?) {}
+    fun <T : BlockEntity> handleTick(level: Level, state: BlockState?, type: BlockEntityType<T>) {}
 }