+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add cross-mod resource - peripheralium
 - Recipe builders for smelting and shapeless recipes
+- BlockEntity support as library part
+-